get current anime season tested successfully in postman

+++ src/main/java/com/anime/AnimeAppApi/jikan/model/seasonNow/Pagination.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class Pagination {
-    public int last_visible_page;
+    public long last_visible_page;
     public boolean has_next_page;
-    public int current_page;
+    public long current_page;
     public Items items;
 }
